Removed unnecessary check in move_to_next_position for CombinationsWithReplacement
When iterating over the `positions` vector to find the position to start
writing the `next_position` from, the current implementation of
`move_to_next_position` for `CombinationsWithReplacement` checks that the
candidate position is either 1) the last position in the vector or 2)
less than the following position. However, `move_to_next_position` only
generates non-decreasing `positions` vectors (and this check does not
contribute to this). Therefore, this check will always pass and can
thus safely be removed.

All tests still pass when running `cargo test`.