feat/permutations: switch to explicit head and one entry per element
In the future, it may be desirable to use a header/footer and get rid of
the `Option<>` wrappers around `prev` and `next`, but the status quo was
a middle ground where the use of `Option<>` made we were checking for
`prev` to be not `None` anyway, so it's easy enough to use an explicit
head, and we were in a painful middle ground.

Additionally, added more doctests for other permutation variants (all
lengths and specified length.

Signed-off-by: Oliver Calder <oliver@calder.dev>